fix : organize the toolkit, add references and reference sources
