+++ hippo4j-core/src/main/java/cn/hippo4j/core/toolkit/inet/InetUtils.java
@@ -35,7 +35,8 @@ import java.util.concurrent.Future;
 import java.util.concurrent.TimeUnit;
 
 /**
- * Inet utils.
+ * Inet utils.<br>
+ * Refer to org.springframework.cloud.commons.util.InetUtils<br>
  */
 public class InetUtils implements Closeable {
 
@@ -80,7 +81,7 @@ public class InetUtils implements Closeable {
                     this.log.trace("Testing interface: " + ifc.getDisplayName());
                     if (ifc.getIndex() < lowest || result == null) {
                         lowest = ifc.getIndex();
-                    } else if (result != null) {
+                    } else {
                         continue;
                     }
                     // @formatter:off
