+++ hippo4j-server/hippo4j-config/src/main/java/cn/hippo4j/config/notify/NotifyCenter.java
@@ -17,12 +17,12 @@
 
 package cn.hippo4j.config.notify;
 
+import cn.hippo4j.common.toolkit.MapUtil;
 import cn.hippo4j.config.event.AbstractEvent;
+import cn.hippo4j.config.event.AbstractSlowEvent;
 import cn.hippo4j.config.notify.listener.AbstractSmartSubscriber;
 import cn.hippo4j.config.notify.listener.AbstractSubscriber;
 import cn.hippo4j.config.toolkit.ClassUtil;
-import cn.hippo4j.config.toolkit.MapUtil;
-import cn.hippo4j.config.event.AbstractSlowEvent;
 import lombok.extern.slf4j.Slf4j;
 
 import java.util.Map;
